requirements.txt
accelerate==0.32.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.3.0
babel==2.16.0
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.12
defusedxml==0.7.1
dill==0.3.8
editdistance==0.8.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
flatbuffers==25.2.10
folium==0.19.4
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.8.0
gradio==4.38.1
gradio_client==1.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.24.0
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcap==1.2.2
mcap-protobuf-support==0.5.2
mdurl==0.1.2
mistune==3.1.0
mmcv==1.4.0
moondream==0.0.5
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
narwhals==1.23.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.5
nuscenes-devkit==1.1.9
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ollama==0.4.7
onnx==1.17.0
onnxruntime==1.21.1
opencv-python==4.11.0.86
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycolmap==3.11.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2025.2
pyvips==2.2.3
pyvips-binary==8.16.0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruff==0.11.8
safetensors==0.5.2
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
semver==3.0.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
stamina==24.3.0
starlette==0.46.2
streamlit==1.45.0
sympy==1.14.0
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.0
triton==3.3.0
typer==0.15.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.127
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
widgetsnbextension==4.0.13
xxhash==3.5.0
xyzservices==2025.1.0
yapf==0.43.0
yarl==1.20.0
